+++ src/packages/form/form.taro.tsx
@@ -1,25 +1,11 @@
-import React, { ReactNode } from 'react'
+import React from 'react'
 import classNames from 'classnames'
-import { Form as TForm, FormProps as TFormProps } from '@tarojs/components'
+import { Form as TForm } from '@tarojs/components'
 import { Context } from './context'
 import { SECRET, useForm } from './useform.taro'
 import { ComponentDefaults } from '@/utils/typings'
 import Cell from '@/packages/cell/index.taro'
-import { FormInstance } from '@/packages/form/types'
-
-export interface FormProps extends TFormProps {
-  footer: ReactNode
-  initialValues: any
-  name: string
-  form: any
-  disabled: boolean
-  divider: boolean
-  validateTrigger: string | string[] | false
-  labelPosition: 'top' | 'left' | 'right'
-  starPosition: 'left' | 'right'
-  onFinish: (values: any) => void
-  onFinishFailed: (values: any, errorFields: any) => void
-}
+import { TaroFormProps, FormInstance } from '@/types'
 
 const defaultProps = {
   ...ComponentDefaults,
@@ -28,9 +14,11 @@ const defaultProps = {
   disabled: false,
   divider: false,
   validateTrigger: 'onChange',
+  onReset: () => {},
+  onSubmit: () => {},
   onFinish: (values) => {},
   onFinishFailed: (values, errorFields) => {},
-} as FormProps
+} as TaroFormProps
 
 const PositionInfo: any = {
   top: 'form-layout-top',
@@ -38,7 +26,7 @@ const PositionInfo: any = {
   right: 'form-layout-right',
 }
 
-export const Form = React.forwardRef<FormInstance, Partial<FormProps>>(
+export const Form = React.forwardRef<FormInstance, Partial<TaroFormProps>>(
   (props, ref) => {
     const classPrefix = 'nut-form'
     const {
@@ -54,6 +42,8 @@ export const Form = React.forwardRef<FormInstance, Partial<FormProps>>(
       validateTrigger,
       labelPosition,
       starPosition,
+      onReset,
+      onSubmit,
       form,
       ...rest
     } = {
@@ -97,11 +87,13 @@ export const Form = React.forwardRef<FormInstance, Partial<FormProps>>(
           e.preventDefault()
           e.stopPropagation()
           submit()
+          onSubmit?.()
         }}
         onReset={(e) => {
           e.preventDefault()
           e.stopPropagation()
           resetFields()
+          onReset?.()
         }}
       >
         <Cell.Group divider={divider}>
