+++ src/packages/form/form.tsx
@@ -1,24 +1,10 @@
-import React, { ReactNode } from 'react'
+import React from 'react'
 import classNames from 'classnames'
 import { Context } from './context'
 import { SECRET, useForm } from './useform'
-import { BasicComponent, ComponentDefaults } from '@/utils/typings'
+import { ComponentDefaults } from '@/utils/typings'
 import Cell from '@/packages/cell'
-import { FormInstance } from '@/packages/form/types'
-
-export interface FormProps extends BasicComponent {
-  footer: ReactNode
-  initialValues: any
-  name: string
-  form: any
-  disabled: boolean
-  divider: boolean
-  validateTrigger: string | string[] | false
-  labelPosition: 'top' | 'left' | 'right'
-  starPosition: 'left' | 'right'
-  onFinish: (values: any) => void
-  onFinishFailed: (values: any, errorFields: any) => void
-}
+import { WebFormProps, FormInstance } from '@/types'
 
 const defaultProps = {
   ...ComponentDefaults,
@@ -27,9 +13,11 @@ const defaultProps = {
   disabled: false,
   divider: false,
   validateTrigger: 'onChange',
+  onReset: () => {},
+  onSubmit: () => {},
   onFinish: (values) => {},
   onFinishFailed: (values, errorFields) => {},
-} as FormProps
+} as WebFormProps
 
 const PositionInfo: any = {
   top: 'form-layout-top',
@@ -37,7 +25,7 @@ const PositionInfo: any = {
   right: 'form-layout-right',
 }
 
-export const Form = React.forwardRef<FormInstance, Partial<FormProps>>(
+export const Form = React.forwardRef<FormInstance, Partial<WebFormProps>>(
   (props, ref) => {
     const classPrefix = 'nut-form'
     const {
@@ -50,6 +38,8 @@ export const Form = React.forwardRef<FormInstance, Partial<FormProps>>(
       disabled,
       onFinish,
       onFinishFailed,
+      onSubmit,
+      onReset,
       validateTrigger,
       labelPosition,
       starPosition,
@@ -94,11 +84,13 @@ export const Form = React.forwardRef<FormInstance, Partial<FormProps>>(
           e.preventDefault()
           e.stopPropagation()
           submit()
+          onSubmit?.()
         }}
         onReset={(e) => {
           e.preventDefault()
           e.stopPropagation()
           resetFields()
+          onReset?.()
         }}
       >
         <Cell.Group divider={divider}>
