+++ src/packages/formitem/index.taro.ts
@@ -1,5 +1,4 @@
 import { FormItem } from './formitem.taro'
 
-export type { FormItemProps } from './formitem.taro'
-export type { FormItemRule } from './types'
+export type { FormItemRule, TaroFormItemProps as FormItemProps } from '@/types'
 export default FormItem
