+++ src/packages/formitem/index.ts
@@ -1,5 +1,4 @@
 import { FormItem } from './formitem'
 
-export type { FormItemProps } from './formitem'
-export type { FormItemRule } from './types'
+export type { FormItemRule, WebFormItemProps as FormItemProps } from '@/types'
 export default FormItem
