+++ src/types/spec/form/h5.ts
@@ -1,3 +1,4 @@
-import { BaseForm } from './base'
+import { BaseForm, BaseFormItem } from './base'
 
 export interface WebFormProps extends BaseForm {}
+export interface WebFormItemProps extends BaseFormItem {}
