+++ src/types/spec/form/taro.ts
@@ -1,3 +1,4 @@
-import { BaseForm } from './base'
+import { BaseForm, BaseFormItem } from './base'
 
 export interface TaroFormProps extends BaseForm {}
+export interface TaroFormItemProps extends BaseFormItem {}
